@@ -1,10 +1,11 @@
 package eu.mobile.onko.activities;
 
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
+
+import androidx.appcompat.app.AppCompatActivity;
 
 import eu.mobile.onko.R;
 import eu.mobile.onko.globalClasses.GlobalData;
