@@ -1,35 +1,31 @@
 package eu.mobile.onko.activities.addMkb;
 
-import android.app.Activity;
 import android.content.Intent;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.view.View;
-import android.view.ViewGroup;
 import android.widget.AdapterView;
 import android.widget.ArrayAdapter;
 import android.widget.ImageView;
 import android.widget.ListView;
 import android.widget.TextView;
 
+import androidx.appcompat.app.AppCompatActivity;
+
 import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
 
 import java.util.ArrayList;
-import java.util.zip.Inflater;
 
 import eu.mobile.onko.R;
-import eu.mobile.onko.adapters.MkbAdapter;
 import eu.mobile.onko.communicationClasses.PostRequest;
 import eu.mobile.onko.communicationClasses.ResponseListener;
 import eu.mobile.onko.globalClasses.GlobalData;
 import eu.mobile.onko.globalClasses.Utils;
 import eu.mobile.onko.models.MkbModel;
 
-public class MkbGroupsActivity extends AppCompatActivity implements View.OnClickListener, AdapterView.OnItemClickListener, ResponseListener {
+public class MkbGroupsActivity extends AppCompatActivity
+        implements View.OnClickListener, AdapterView.OnItemClickListener, ResponseListener {
 
     private static final int        REQUEST_CODE_MKB    = 0;
 
