@@ -4,6 +4,11 @@ import android.content.Context;
 import android.graphics.Typeface;
 import android.widget.TextView;
 
+import java.text.ParseException;
+import java.text.SimpleDateFormat;
+import java.util.Date;
+import java.util.Locale;
+
 /**
  * Created by stoycho.petrov on 05/09/2017.
  */
@@ -11,7 +16,7 @@ import android.widget.TextView;
 public class Utils {
 
     // REQUEST LINKS
-    public static final String URL                      = "http://90.154.211.179:50110/api/";
+    public static final String URL                      = "http://192.168.1.7:59181/api/";
     public static final String LOGIN                    = "users/login";
     public static final String REGISTER_PUSH_TOKEN      = "users/registerPushToken";
     public static final String REGISTER                 = "users/register";
@@ -21,10 +26,24 @@ public class Utils {
     public static final String ADD_ACTION               = "/add";
     public static final String DELETE_ACTION            = "/delete";
     public static final String USER_MKB_CONTROLLER      = "userMkb";
+    public static final String GET_SCHEDULERS           = "schedulers/getUserScheduler";
+    public static final String SCHEDULES                = "Schedules/";
     public static final String GET_EXAMINATIONS         = "examinations/getByMkb";
+    public static final String GET_EXAMINATION          = "examinations";
     public static final String GET_DOCTORS              = "doctors/getDoctors";
+    public static final String GET_USERS_DOCTORS        = "users/doctors?doctorId=";
     public static final String GET_USER_EXAMINATION     = "userExaminations/getUserExaminations";
     public static final String EXAMINATIONS_CONTROLLER  = "userExaminations";
+    public static final String DOCTORS_LIST             = "doctors/getDoctorsList";
+    public static final String GET_FREE_HOURS           = "schedulers/getDoctorsFreeHours";
+    public static final String RESERVATIONS             = "reservations/";
+    public static final String GET_RESERVATIONS_BY_ID   = "getReservationsByClient";
+    public static final String GET_RESERVATIONS_BY_DOCTOR = "getReservationsByDoctor";
+
+    public static final String DATE_FORMAT              = "dd/MM/yyyy";
+    public static final String SERVER_DATE_FORMAT       = "yyyy-MM-dd";
+    public static final String TIME_FORMAT              = "HH:mm";
+    public static final String SERVER_DATE_TIME         = "yyyy-MM-dd HH:mm:ss";
 
     // QUERY PARAMS
     public static final String GROUP_ID     = "?groupId=";
@@ -48,6 +67,8 @@ public class Utils {
     public static final String USER_FIRST_NAME  = "user_first_name";
     public static final String USER_LAST_NAME   = "user_last_name";
     public static final String USER_PHONE       = "user_phone";
+    public static final String USER_DOCTOR_ID   = "doctor_id";
+    public static final String USER_TYPE_ID     = "user_type_id";
     public static final String USER_MKB         = "UserMKB";
     public static final String USER_TOKEN       = "auth_token";
     public static final String MKB_ID           = "MkbId";
@@ -58,6 +79,7 @@ public class Utils {
     public static final String USER_MKB_ID      = "UserMkbId";
     public static final String EXAMINATION_ID   = "examination_id";
     public static final String EXAMINATION_NAME = "examination_name";
+    public static final String EXAMINATION      = "examination";
     public static final String MKB_GROUP        = "mkb_group";
     public static final String DOCTORS          = "doctors";
     public static final String DOCTOR_NAME      = "doctor_name";
@@ -73,12 +95,24 @@ public class Utils {
     public static final String EXAMINATION_DATE = "examination_date";
     public static final String USER_EXAMINATION_ID  = "user_examination_id";
     public static final String PUSH_TOKEN           = "push_token";
+    public static final String DOCTOR_ID            = "doctor_id";
+    public static final String SCHEDULE_ID          = "schedule_id";
+    public static final String START_HOUR           = "start_hour";
+    public static final String END_HOUR             = "end_hour";
+    public static final String WEEK_DAY_ID          = "week_day_id";
+    public static final String USER_ID              = "user_id";
+    public static final String FREE_HOURS           = "free_hours";
+    public static final String CLIENT_ID            = "client_id";
+    public static final String IS_COMPLETED         = "is_completed";
 
 
     // INTENT EXTRAS
     public static final String INTENT_GROUP_ID      = "group_id";
     public static final String INTENT_GROUP_NAME    = "group_name";
     public static final String INTENT_MKB_ID        = "mkb_id";
+    public static final String INTENT_DOCTOR        = "doctor";
+    public static final String INTENT_DOCTOR_ID     = "doctor_id";
+    public static final String INTENT_EXAMINATION_ID = "examination_id";
 
     // PREFERENCES
     public static final String  PREFERENCES_PUSH_TOKEN                          = "push_token";
@@ -97,5 +131,33 @@ public class Utils {
     public static void setTypeFace(Context context, TextView textView, String typeFacePath){
         Typeface customFont = Typeface.createFromAsset(context.getAssets(),  typeFacePath);
         textView.setTypeface(customFont);
+    }
+
+    public static String getAccountTypeName(Context context, int type) {
+        try {
+            String name = String.format("account_type_%1s", type);
+            int flagResid = context.getResources().getIdentifier(name, "string", context.getPackageName());
+            return context.getString(flagResid);
+        }catch (Exception exception) {
+            exception.printStackTrace();
+        }
+
+        return "";
+    }
+
+    public static String formatDate(Date date, String dateFormat) {
+        SimpleDateFormat simpleDateFormat = new SimpleDateFormat(dateFormat, Locale.getDefault());
+        return simpleDateFormat.format(date);
+    }
+
+    public static Date parseDate(String date) {
+        try {
+            SimpleDateFormat simpleDateFormat = new SimpleDateFormat(DATE_FORMAT, Locale.getDefault());
+            return simpleDateFormat.parse(date);
+        } catch (ParseException e) {
+            e.printStackTrace();
+        }
+
+        return null;
     }
 }
