@@ -1,10 +1,13 @@
 package eu.mobile.onko.models;
 
+import android.os.Parcel;
+import android.os.Parcelable;
+
 /**
  * Created by Stoycho Petrov on 15.5.2018 г..
  */
 
-public class DoctorModel {
+public class DoctorModel implements Parcelable {
 
     private int         mDoctorId;
     private String      mDoctorName;
@@ -13,6 +16,23 @@ public class DoctorModel {
         this.mDoctorId      = mDoctorId;
         this.mDoctorName    = mDoctorName;
     }
+
+    protected DoctorModel(Parcel in) {
+        mDoctorId = in.readInt();
+        mDoctorName = in.readString();
+    }
+
+    public static final Creator<DoctorModel> CREATOR = new Creator<DoctorModel>() {
+        @Override
+        public DoctorModel createFromParcel(Parcel in) {
+            return new DoctorModel(in);
+        }
+
+        @Override
+        public DoctorModel[] newArray(int size) {
+            return new DoctorModel[size];
+        }
+    };
 
     public int getmDoctorId() {
         return mDoctorId;
@@ -28,5 +48,16 @@ public class DoctorModel {
 
     public void setmDoctorName(String mDoctorName) {
         this.mDoctorName = mDoctorName;
+    }
+
+    @Override
+    public int describeContents() {
+        return 0;
+    }
+
+    @Override
+    public void writeToParcel(Parcel dest, int flags) {
+        dest.writeInt(mDoctorId);
+        dest.writeString(mDoctorName);
     }
 }
